Initial pnpm integration
