+++ client/auth/package.json
@@ -1,0 +1,5 @@
+{
+  "private": true,
+  "name": "@client/auth",
+  "version": "1.0.0"
+}
