+++ client/orchest/package.json
@@ -1,0 +1,5 @@
+{
+  "private": true,
+  "name": "@client/orchest",
+  "version": "1.0.0"
+}
