+++ lib/javascript/mdc-components/package.json
@@ -1,0 +1,37 @@
+{
+  "private": true,
+  "name": "@lib/mdc",
+  "version": "1.0.0",
+  "scripts": {
+    "bundle": "esbuild src/index.js --bundle --loader:.js=jsx --outfile=dist/index.js --define:process.env.NODE_ENV=\\\"development\\\" --define:global=window --sourcemap --external:lib-utils --external:react*",
+    "build": "pnpm run bundle -- --target=chrome58,firefox57,safari11,edge16",
+    "build-development": "pnpm run bundle"
+  },
+  "peerDependencies": {
+    "react": ">= 16.13.1",
+    "react-dom": ">= 16.13.1"
+  },
+  "dependencies": {
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "@material/checkbox": "10.0.0",
+    "@material/circular-progress": "10.0.0",
+    "@material/data-table": "10.0.0",
+    "@material/dialog": "10.0.0",
+    "@material/form-field": "10.0.0",
+    "@material/icon-button": "10.0.0",
+    "@material/linear-progress": "10.0.0",
+    "@material/radio": "10.0.0",
+    "@material/ripple": "10.0.0",
+    "@material/select": "10.0.0",
+    "@material/switch": "10.0.0",
+    "@material/tab-bar": "10.0.0",
+    "@material/textfield": "10.0.0",
+    "@lib/utils": "workspace:*"
+  },
+  "devDependencies": {
+    "esbuild": "^0.8.44",
+    "react": "16.13.1",
+    "react-dom": "16.13.1"
+  }
+}
