+++ lib/javascript/mdc-components/src/MDCButtonReact.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from "react";
 import { MDCRipple } from "@material/ripple";
-import { RefManager } from "../utils/all";
+import { RefManager } from "@lib/utils";
 
 // used in orchest-webserver and orchest-authserver
 class MDCButtonReact extends React.Component {
