+++ lib/javascript/mdc-components/src/MDCCheckboxReact.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { MDCFormField } from "@material/form-field";
 import { MDCCheckbox } from "@material/checkbox";
-import { RefManager, uuidv4 } from "../utils/all";
+import { RefManager, uuidv4 } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCCheckboxReact extends React.Component {
