+++ lib/javascript/mdc-components/src/MDCCircularProgressReact.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { MDCCircularProgress } from "@material/circular-progress";
-import { RefManager } from "../utils/all";
+import { RefManager } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCCircularProgressReact extends React.Component {
