+++ lib/javascript/mdc-components/src/MDCDataTableReact.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { MDCDataTable } from "@material/data-table";
-import { RefManager, someParentHasClass } from "../utils/all";
+import { RefManager, someParentHasClass } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCDataTableReact extends React.Component {
