+++ lib/javascript/mdc-components/src/MDCDialogReact.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { MDCDialog } from "@material/dialog";
-import { RefManager, uuidv4 } from "../utils/all";
+import { RefManager, uuidv4 } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCDialogReact extends React.Component {
