+++ lib/javascript/mdc-components/src/MDCIconButtonToggleReact.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { MDCIconButtonToggle } from "@material/icon-button";
-import { RefManager } from "../utils/all";
+import { RefManager } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCIconButtonToggleReact extends React.Component {
