+++ lib/javascript/mdc-components/src/MDCLinearProgressReact.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { MDCLinearProgress } from "@material/linear-progress";
-import { RefManager } from "../utils/all";
+import { RefManager } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCLinearProgressReact extends React.Component {
