+++ lib/javascript/mdc-components/src/MDCRadioReact.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { MDCFormField } from "@material/form-field";
 import { MDCRadio } from "@material/radio";
-import { RefManager, uuidv4 } from "../utils/all";
+import { RefManager, uuidv4 } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCRadioReact extends React.Component {
