+++ lib/javascript/mdc-components/src/MDCSelectReact.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { MDCSelect } from "@material/select";
-import { arraysEqual, RefManager, uuidv4 } from "../utils/all";
+import { arraysEqual, RefManager, uuidv4 } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCSelectReact extends React.Component {
