+++ lib/javascript/mdc-components/src/MDCSwitchReact.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { MDCSwitch } from "@material/switch";
-import { RefManager, uuidv4 } from "../utils/all";
+import { RefManager, uuidv4 } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCSwitchReact extends React.Component {
