+++ lib/javascript/mdc-components/src/MDCTabBarReact.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { MDCTabBar } from "@material/tab-bar";
-import { RefManager } from "../utils/all";
+import { RefManager } from "@lib/utils";
 
 // used only in orchest-webserver
 class MDCTabBarReact extends React.Component {
