+++ lib/javascript/mdc-components/src/MDCTextFieldReact.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from "react";
 import { MDCTextField } from "@material/textfield";
-import { RefManager, uuidv4 } from "../utils/all";
+import { RefManager, uuidv4 } from "@lib/utils";
 
 // used in orchest-webserver and orchest-authserver
 class MDCTextFieldReact extends React.Component {
