+++ lib/javascript/mdc-components/src/index.js
@@ -1,0 +1,12 @@
+export * from "./MDCButtonReact";
+export * from "./MDCCheckboxReact";
+export * from "./MDCCircularProgressReact";
+export * from "./MDCDataTableReact";
+export * from "./MDCDialogReact";
+export * from "./MDCIconButtonToggleReact";
+export * from "./MDCLinearProgressReact";
+export * from "./MDCRadioReact";
+export * from "./MDCSelectReact";
+export * from "./MDCSwitchReact";
+export * from "./MDCTabBarReact";
+export * from "./MDCTextFieldReact";
