+++ lib/javascript/utils/package.json
@@ -1,0 +1,20 @@
+{
+  "private": true,
+  "name": "@lib/utils",
+  "version": "1.0.0",
+  "main": "src/index.js",
+  "scripts": {
+    "bundle": "esbuild src/index.js --bundle --loader:.js=jsx --outfile=dist/index.js --define:process.env.NODE_ENV=\\\"development\\\" --define:global=window --sourcemap --external:react*",
+    "build": "pnpm run bundle -- --target=chrome58,firefox57,safari11,edge16",
+    "build-development": "pnpm run bundle"
+  },
+  "peerDependencies": {
+    "react": ">= 16.13.1",
+    "react-dom": ">= 16.13.1"
+  },
+  "devDependencies": {
+    "esbuild": "^0.8.44",
+    "react": "16.13.1",
+    "react-dom": "16.13.1"
+  }
+}
