+++ pnpm-workspace.yaml
@@ -1,0 +1,7 @@
+packages:
+  # all packages in subdirs of packages/ and components/
+  - "client/**"
+  - "lib/javascript/**"
+  - "services/**"
+  # exclude packages that are inside test directories
+  - "!**/test/**"
