+++ services/orchest-webserver/app/static/js/src/utils/webserver-utils.js
@@ -226,29 +226,6 @@ export function cleanServerDateTime(dateTimeString) {
   return dateTimeString.replace(regex, subst);
 }
 
-export function newslines2breaks(lines) {
-  if (lines === undefined) {
-    return [];
-  }
-
-  // subtitute newlines for line breaks
-  let linesArr = lines.split("\n");
-
-  let lineElements = linesArr.map((line, index) => {
-    if (index !== linesArr.length - 1) {
-      return (
-        <Fragment key={index}>
-          {line}
-          <br />
-        </Fragment>
-      );
-    } else {
-      return <Fragment key={index}>{line}</Fragment>;
-    }
-  });
-  return lineElements;
-}
-
 export function getPipelineJSONEndpoint(
   pipeline_uuid,
   project_uuid,
